some changes about design

+++ database/migrations/2020_03_26_173633_create_labours_table.php
@@ -24,7 +24,7 @@ class CreateLaboursTable extends Migration
             $table->integer('food_rate');
             $table->bigInteger('total_food_rate')->default(0);
             $table->bigInteger('due_foodrate')->default(0);
-            $table->bigInteger('total_attendance')->default(0);
+            $table->decimal('total_attendance', 8, 2)->default(0);
             $table->bigInteger('total_salary')->default(0);
             $table->bigInteger('total_paid')->default(0);
             $table->bigInteger('total_due')->default(0);
